Fix 64-bit build. Avoid using variable-sized types in Atlas message interface.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9251 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/GameInterface/Handlers/MapHandlers.cpp
@@ -86,10 +86,10 @@ QUERYHANDLER(GenerateMap)
 
 	CScriptValRooted settings;
 	scriptInterface.Eval("({})", settings);
-	scriptInterface.SetProperty(settings.get(), "Size", (size_t)msg->size);
-	scriptInterface.SetProperty(settings.get(), "Seed", (size_t)msg->seed);
+	scriptInterface.SetProperty(settings.get(), "Size", (int)msg->size);
+	scriptInterface.SetProperty(settings.get(), "Seed", (int)msg->seed);
 	scriptInterface.SetProperty(settings.get(), "BaseTerrain", std::vector<std::wstring>(*msg->terrain));
-	scriptInterface.SetProperty(settings.get(), "BaseHeight", (size_t)msg->height);
+	scriptInterface.SetProperty(settings.get(), "BaseHeight", (int)msg->height);
 	scriptInterface.SetProperty(settings.get(), "CircularMap", true);	// now default to circular map
 
 	CScriptValRooted pData = scriptInterface.ParseJSON(*msg->playerData);
